backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1

psycopg2-binary
celery==5.4.0
redis==5.0.4
python-multipart==0.0.9
pydantic-settings==2.2.1
aiofiles==23.2.1
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==1.1.0